refactor wrangler result section to separate component
