+++ cdap-ui/app/cdap/components/Wrangler/WrangleHistory/index.js
@@ -1,0 +1,37 @@
+/*
+ * Copyright © 2016 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import React, {PropTypes} from 'react';
+
+export default function WrangleHistory({historyArray}) {
+  if (historyArray.length === 0) { return null; }
+
+  return (
+    <ul>
+      {
+        historyArray.map((history) => {
+          return (
+            <li key={history.id}>{history.action}: {history.payload.join(', ')}</li>
+          );
+        })
+      }
+    </ul>
+  );
+}
+
+WrangleHistory.propTypes = {
+  historyArray: PropTypes.array
+};
